aurum_backend/requirements.txt
aiofiles==24.1.0
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
Deprecated==1.2.18
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-spectacular==0.28.0
ecdsa==0.19.1
email-validator==2.0.0.post2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.103.2
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
fonttools==4.58.5
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
limits==4.2
makefun==1.16.0
MarkupSafe==3.0.2
matplotlib==3.9.4
narwhals==1.45.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.2
pandas==2.1.3
passlib==1.7.4
pdfkit==1.0.0
pillow==11.3.0
plotly==6.2.0
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.8.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
schedule==1.2.2
seaborn==0.13.2
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlparse==0.5.3
starlette==0.27.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.1.0
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
zipp==3.23.0
zopfli==0.2.3.post1
